requirements.txt
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
fastapi==0.112.0
filelock==3.15.4
frozenlist==1.4.1
googleapis-common-protos==1.59.1
grpcio==1.65.4
h11==0.14.0
idna==3.7
importlib_metadata==8.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
msgpack==1.0.8
opentelemetry-api==1.26.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-requests==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
packaging==24.1
protobuf==4.25.4
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.1
ray==2.34.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
thrift==0.20.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
wrapt==1.16.0
zipp==3.19.2